speed
Former-commit-id: e3b719c706c23f89886ac07d4b75dc67a0c24903

+++ kevoree-tools/org.kevoree.tools.aether.framework.android/src/main/scala/org/kevoree/tools/aether/framework/android/AndroidKevoreeJarClassLoader.scala
@@ -13,12 +13,12 @@
  */
 package org.kevoree.tools.aether.framework.android
 
-import dalvik.system.DexClassLoader
 import android.content.Context
 import java.io.{BufferedOutputStream, FileOutputStream, File, InputStream}
 import org.kevoree.kcl.{KevoreeResourcesLoader, KevoreeJarClassLoader}
 import org.slf4j.LoggerFactory
 import java.lang.Class
+import dalvik.system.{DexFile, DexClassLoader}
 
 /**
  * Created by IntelliJ IDEA.
@@ -35,10 +35,16 @@ class AndroidKevoreeJarClassLoader(gkey: String, ctx: android.content.Context, p
 
   val selfPointer = this
 
+
+
+
   private class KevoreeDexClassLoader(c1:String,c2:String,c3:String,parentCL : ClassLoader) extends DexClassLoader(c1,c2,c3,parentCL) {
+
+   // val dexFile = DexFile.loadDex(c1,c2,0)
 
     def internalLoad(name:String) : Class[_] = {
       super[DexClassLoader].loadClass(name)
+      //dexFile.loadClass(name,this)
     }
 
     override def loadClass(p1: String) : Class[_] = {
@@ -46,6 +52,9 @@ class AndroidKevoreeJarClassLoader(gkey: String, ctx: android.content.Context, p
     }
 
   }
+
+
+
 
 
   /* Constructor */
@@ -67,21 +76,32 @@ class AndroidKevoreeJarClassLoader(gkey: String, ctx: android.content.Context, p
 
   def declareLocalDexClassLoader(dexStream: InputStream, idName: String) {
     logger.debug("Begin declare subClassLoader " + idName)
-    val cleanName = System.currentTimeMillis() + idName.replaceAll(File.separator, "_").replaceAll(":", "_")
+    //val cleanName = System.currentTimeMillis() + idName.replaceAll(File.separator, "_").replaceAll(":", "_")
+
+    val cleanName : String = (if(idName.contains("SNAPSHOT")){
+      System.currentTimeMillis() + idName.replaceAll(File.separator, "_").replaceAll(":", "_")
+    } else {
+      idName.replaceAll(File.separator, "_").replaceAll(":", "_")
+    })
+
     val dexInternalStoragePath = new File(ctx.getDir("dex", Context.MODE_WORLD_WRITEABLE), cleanName)
-    logger.debug("File Create " + dexInternalStoragePath.getAbsolutePath)
-    val dexWriter = new BufferedOutputStream(new FileOutputStream(dexInternalStoragePath))
-    val b = new Array[Byte](2048)
-    var len = 0;
-    while (len != -1) {
-      len = dexStream.read(b);
-      if (len > 0) {
-        dexWriter.write(b, 0, len);
+    if (!dexInternalStoragePath.exists()){
+      logger.debug("File Create " + dexInternalStoragePath.getAbsolutePath)
+      val dexWriter = new BufferedOutputStream(new FileOutputStream(dexInternalStoragePath))
+      val b = new Array[Byte](2048)
+      var len = 0;
+      while (len != -1) {
+        len = dexStream.read(b);
+        if (len > 0) {
+          dexWriter.write(b, 0, len);
+        }
       }
+      dexWriter.flush()
+      dexWriter.close()
     }
-    dexWriter.flush()
-    dexWriter.close()
-    val dexOptStoragePath = ctx.getDir("odex" + System.currentTimeMillis(), Context.MODE_WORLD_WRITEABLE)
+
+    val dexOptStoragePath = ctx.getDir("odex" + cleanName, Context.MODE_WORLD_WRITEABLE)
+    //logger.info("create ODEX dir = "+dexOptStoragePath.mkdirs()+"-"+dexOptStoragePath.isDirectory)
     odexPaths = odexPaths ++ List(dexOptStoragePath)
     dexOptStoragePath.mkdirs()
     val newDexCL = new KevoreeDexClassLoader(dexInternalStoragePath.getAbsolutePath, dexOptStoragePath.getAbsolutePath, null, parent)
